Refactor player session management to implement a simplified player session approach. This update replaces the complex `usePlayerSessions` hook with a streamlined version, modifies the `PlayerSessionsProvider` and `useHomePage` to utilize the new interface, and enhances the HomePage UI to reflect automatic player simulation controls. The `useSimplifiedPlayerSession` hook has been removed, and all integration points have been updated accordingly.
